New API, inMemorySocketPair()
I'm using Array<Socket> instead of Pair<Socket, Socket>
because Array is _slightly_ more friendly to Java-language
consumers.
